Seperated auth and tested to work
