+++ TravelExperts/Views/Account/_Layout.cshtml
@@ -39,7 +39,7 @@
                 </li>
             </ul>
             <div class="sidebar-footer">
-                <form asp-action="Logout" method="post" class="d-inline">
+                <form asp-controller="Auth" asp-action="Logout" method="post" class="d-inline">
                     <button type="submit" class="btn sidebar-link">
                         <i class="lni lni-exit"></i>
                         <span>Logout</span>
