+++ TravelExperts/Views/Shared/_LoginPartial.cshtml
@@ -7,7 +7,7 @@
                         <a class="nav-link text-dark" asp-controller="Account" asp-action="Index">Welcome, @User.Identity.Name!</a>
                     </li>
                     <li class="nav-item">
-                        <form asp-controller="Account" asp-action="Logout" method="post">
+                        <form asp-controller="Auth" asp-action="Logout" method="post">
                             <button id="logout" type="submit" class="nav-link btn btn-link text-dark border-0">Logout</button>
                         </form>
                     </li>
@@ -17,10 +17,10 @@
             {
                 <ul class="navbar-nav">
                     <li class="nav-item">
-                        <a class="nav-link text-dark" asp-controller="Account" asp-action="Login">Login</a>
+                        <a class="nav-link text-dark" asp-controller="Auth" asp-action="Login">Login</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-dark" asp-controller="Account" asp-action="Register">Register</a>
+                        <a class="nav-link text-dark" asp-controller="Auth" asp-action="Register">Register</a>
                     </li>
                 </ul>
             }
